+++ packages/indexer/src/index.ts
@@ -1,6 +1,7 @@
 import { config as dotEnvConfig } from "dotenv";
 dotEnvConfig();
 
+import "@/jobs/cdc/index";
 import "@/common/tracer";
 import "@/config/polyfills";
 import "@/jobs/index";
@@ -12,7 +13,7 @@ import { config } from "@/config/index";
 import { logger } from "@/common/logger";
 import { getNetworkSettings } from "@/config/network";
 import { Sources } from "@/models/sources";
-import { startKafkaConsumer, startKafkaProducer } from "./jobs/cdc";
+import { startKafkaConsumer, startKafkaProducer } from "@/jobs/cdc/index";
 
 process.on("unhandledRejection", (error) => {
   logger.error("process", `Unhandled rejection: ${error}`);
@@ -32,8 +33,10 @@ const setup = async () => {
   }
 
   if (config.doKafkaWork) {
-    await startKafkaConsumer();
-    await startKafkaProducer();
+    // eslint-disable-next-line no-console
+    console.log("doKafkaWork");
+    startKafkaConsumer();
+    startKafkaProducer();
   }
 
   if (config) await Sources.getInstance();
